[REFACTOR] : ♻️ MyRequestTab interface 통합
