--- src/components/MyRequest/MyRequestList/interfaces.ts
@@ -1,14 +1,0 @@
-interface MyListProps {
-  eventId: number;
-  userName: string;
-  title: string;
-  file?: string;
-  fileName?: string;
-  startDay: string;
-  endDay: string;
-  status: 'submit' | 'accept' | 'deny';
-}
-
-export interface FileProps {
-  file: MyListProps;
-}
