+++ src/components/MyRequest/interfaces.ts
@@ -1,1 +1,18 @@
-export {};
+interface File {
+  fileName: string;
+  fileLocation: string;
+}
+
+export interface MyListProps {
+  Id: number;
+  userName: string;
+  title: string;
+  files?: File[];
+  start: string;
+  end: string;
+  status: 'submit' | 'accept' | 'deny';
+}
+
+export interface FileProps {
+  file: MyListProps;
+}
